requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
mysql-connector-python-rf==2.2.2
python-dotenv==1.0.1
ujson==5.10.0
Werkzeug==3.1.3
